feat(result): adds support to the results pattern and the semantic returns to the client
When requesting to shorten an URL or to be redirected to an URL, the application returns a
'Result<T>'. If the 'Result' is successful - i.e. contains a value -, the API returns a 2xx-3xx
status code. If the 'Result' is not successful, the ResultsHelper process the error and returns a
specific status code and a detailed description of the error.

#1